singlesource/azure-monitor-docs/tools2/requirements.txt
azure-identity
azure-mgmt-resourcegraph
jinja2
